Refactor model structure to use Traits for BaseModel and AsHasMany
