+++ app/Models/Email.php
@@ -4,12 +4,14 @@ declare(strict_types = 1);
 
 namespace App\Models;
 
-use CodeFusion\Model\BaseModel;
+use CodeFusion\Model\Traits\BaseModel;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
 
-class Email extends BaseModel
+class Email extends Model
 {
     use HasFactory;
+    use BaseModel;
 
     protected $fillable = ['value'];
 }
