+++ src/Model/Traits/BaseModel.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace CodeFusion\Model\Traits;
+
+trait BaseModel
+{
+    use AsHasMany;
+    use Filterable;
+}
